# mvcc: cache consistent index