Called on every entry apply and boltdb requests aren't free.

Old:
```sh
$ go test -bench=Consistent -v -count=5 -run=Consistent
BenchmarkConsistentIndex-8       3000000               452 ns/op              96 B/op          3 allocs/op
BenchmarkConsistentIndex-8       3000000               457 ns/op              96 B/op          3 allocs/op
BenchmarkConsistentIndex-8       3000000               452 ns/op              96 B/op          3 allocs/op
BenchmarkConsistentIndex-8       3000000               466 ns/op              96 B/op          3 allocs/op
BenchmarkConsistentIndex-8       3000000               446 ns/op              96 B/op          3 allocs/op
PASS
ok      _/home/anthony/go/src/github.com/coreos/etcd/mvcc       12.831s
```
New:
```sh
$ GOPATH=~/etcd-gopath/ go test -bench=Consistent -v -count=5 -run=Consistent
BenchmarkConsistentIndex-8      300000000                5.22 ns/op            0 B/op          0 allocs/op
BenchmarkConsistentIndex-8      300000000                5.30 ns/op            0 B/op          0 allocs/op
BenchmarkConsistentIndex-8      300000000                5.22 ns/op            0 B/op          0 allocs/op
BenchmarkConsistentIndex-8      300000000                5.24 ns/op            0 B/op          0 allocs/op
BenchmarkConsistentIndex-8      300000000                5.38 ns/op            0 B/op          0 allocs/op
```